backend/pyproject.toml
[project]
name = "backend"
version = "1.0.0"
description = "FastAPI backend for Echo App"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.3",
    "python-dotenv>=1.0.1",
    "azure-cosmos>=4.9.0",
    "azure-identity>=1.15.0",
    "pyjwt[crypto]>=2.8.0",
    "cachetools>=5.3.0",
    "httpx>=0.27.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "cryptography>=42.0.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
